Small Update #3
This update fix the bad syncing of the radar sound and solve a few minor glitches. It also add a special menu that will get accessible in a futur update (Not now because the tool isn't fully implemented yet. I still need help for a few technical stuff so if you are an expert with unity, contact me on discord @blasterholobot).